Add logic for all remaining funcitons.
Fixed issue with  mod_inv, turns out was not an
issue with the function itself. It was a bad call
located in bezout_coeffs. It was returning a local
variable inside a loop which occasionally would
cause issues.

Additionally, finished up remaining function. It
occasionally raises a ZeroDivisionError which I
need to investigate futher. Unknown what values
input for a, m cause the issue.