QR factorization
working on constructing R matrix which is an upper trangular matrix.
entries are defined as r_{ij} = <u_i, a_j> for i<=j 0 for i > j.

looking at it initially, my r_{ij} entry looks like it's just going to
be the inner product of the i-th row of 'Q' and the j-th col of the
input matrix 'A', so will attempt based that.